General redesign of KDL::CoDyCo::TreeGraph data structure using indeces instead of maps, change joints into junctions to avoid confusion with upstream KDL concept, added junctions serialization

+++ src/treeidserialsolver_recursive_newton_euler.cpp
@@ -73,7 +73,7 @@ namespace CoDyCo{
         int l;
 
         //Sweep from root to leaf
-        for( l = 0; l < recursion_order.size(); l++ ) {
+        for( l = 0; l < (int)recursion_order.size(); l++ ) {
             
             unsigned int curr_index = recursion_order[l];
